refactor: migrate CLI command files into 'commands' module directory
