+++ src/commands/index.ts
@@ -1,0 +1,7 @@
+export * from "./add";
+export * from "./check";
+export * from "./copy";
+export * from "./installations";
+export * from "./publish";
+export * from "./remove";
+export * from "./update";
